refactoring to remove duplication
