+++ src/io/iohk/scalanet/peergroup/UDPPeerGroup.scala
@@ -15,6 +15,7 @@ import monix.eval.Task
 
 import scala.language.higherKinds
 import UDPPeerGroup._
+import io.iohk.decco.PartialCodec
 
 class UDPPeerGroup[F[_]](val config: Config)(implicit liftF: Lift[F])
     extends TerminalPeerGroup[InetSocketAddress, F]() {
@@ -63,6 +64,8 @@ class UDPPeerGroup[F[_]](val config: Config)(implicit liftF: Lift[F])
   }
 
   override val processAddress: InetSocketAddress = config.processAddress
+  override def createMessageChannel[MessageType: PartialCodec](): MessageChannel[InetSocketAddress, MessageType, F] =
+    new TerminalGroupMessageChannel(this, messageStream)
 }
 
 object UDPPeerGroup {
